Rename erc20TokenApproval file
